Move path utils from plugin to shared for reuse
