+++ src/plugin/tests/identifier-case-assets.integration.test.js
@@ -13,7 +13,7 @@ import {
     setIdentifierCaseDryRunContext
 } from "../src/identifier-case/identifier-case-context.js";
 import { prepareIdentifierCasePlan } from "../src/identifier-case/local-plan.js";
-import { fromPosixPath } from "../src/utils/path-utils.js";
+import { fromPosixPath } from "../../shared/path-utils.js";
 
 const currentDirectory = fileURLToPath(new URL(".", import.meta.url));
 const pluginPath = path.resolve(currentDirectory, "../src/gml.js");
